🧬 Move the object about the base class for template feature to module *model.api_config.template*.
