+++ pymock_api/model/api_config/__init__.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, List, Optional, Union
 
 from ..._utils import YAML
 from ..._utils.file_opt import _BaseFileOperation
-from ._base import _Config, _TemplatableConfig
+from ._base import _Config
 from .apis import (
     HTTP,
     APIParameter,
@@ -18,7 +18,7 @@ from .apis import (
 )
 from .base import BaseConfig
 from .item import IteratorItem
-from .template import TemplateConfig, TemplateConfigLoadable
+from .template import TemplateConfig, TemplateConfigLoadable, _TemplatableConfig
 
 
 class MockAPIs(_Config, TemplateConfigLoadable):
