+++ pymock_api/model/api_config/apis/request.py
@@ -1,8 +1,9 @@
 from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional
 
-from .._base import _Config, _TemplatableConfig
+from .._base import _Config
 from ..item import IteratorItem
+from ..template import _TemplatableConfig
 
 
 @dataclass(eq=False)
